refactor: change validator file name to be more generic
