+++ lib/internals/validateObject.ts
