+++ lib/pipes/validators/validation.pipe.ts
@@ -1,6 +1,6 @@
 import type { ClassTransformOptions } from 'class-transformer';
 import type { ValidatorOptions } from 'class-validator';
-import { validateObject } from '../../internals/classValidator';
+import { validateObject } from '../../internals/validateObject';
 import type { ParameterPipe, PipeMetadata } from '../ParameterPipe';
 
 export interface ValidationPipeOptions extends ValidatorOptions {
